add getGenre and  musicBrainz apis to fetch the genres
